Place S3 classes in S3 common module
