+++ bin/deploy
@@ -55,13 +55,13 @@ app_bucket = apps.detect { |app| app.key == app_names[app_index] }
 application.bucket = app_bucket
 puts "App \"#{application.name}\" selected."
 
-require_relative '../deploy/s3_configuration'
-s3 = S3Configuration.new(app_bucket)
 require_relative '../deploy/eb/configuration'
 eb = Eb::Configuration.new(app_bucket)
+require_relative '../deploy/s3/configuration'
+s3 = S3::Configuration.new(app_bucket)
 
-require_relative '../deploy/deployable/s3_platform'
 require_relative '../deploy/eb/platform'
+require_relative '../deploy/s3/platform'
 # TODO: Don't call out to bash, just call local classes to deploy
 if eb.exists?
   application.platform = Eb::Platform.new
@@ -69,7 +69,7 @@ if eb.exists?
   puts "Environment \'#{application.name}\' found on EB."
   puts "You've asked to deploy version \'#{TAG}\'."
 elsif s3.exists?
-  application.platform = S3Platform.new
+  application.platform = S3::Platform.new
   application.platform_config = s3
   puts "Website \'#{application.name}\' found on S3."
   puts "Config version \"#{s3.version}\" selected."
