--- deploy/deployable/s3_platform.rb
@@ -1,20 +1,0 @@
-class S3Platform
-  def deploy(application)
-    @application = application
-    @tag = @application.tag
-    s3_deploy!
-  end
-
-  private
-  def s3
-    @s3 ||= @application.platform_config
-  end
-
-  def s3_deploy!
-    system(
-      "bucket=#{s3.target_bucket}"\
-      " s3_config_version=#{s3.version}"\
-      " npm run publish"
-    )
-  end
-end
